Added image, title and ingredient to RecipeDetailsFragment.kt

+++ app/src/main/java/com/vk/recipebook/dataSources/ApiRecipe/Responses.kt
@@ -29,6 +29,6 @@ data class ResponsesFromRecipeDetails(
 
 data class Ingredients(
     var name: String,
-    var amount: Int,
+    var amount: Float,
     var unit: String
 )
